boom: all keras2 configs importable
